Organized controllers, moved methods with ResponseEntity to a @RestController
