+++ mgl-webapp_task2-0/src/main/java/com/organization/mvcproject/service/GameService.java
@@ -6,7 +6,7 @@ import com.organization.mvcproject.model.Game;
 
 public interface GameService {
 
-	List<Game> retrieveAllGames();
+	List<Game> getAllGames();
 
 	Game saveGame(Game game);
